+++ src/KPCOS.BusinessLayer/DTOs/Response/Projects/ProjectResponse.cs
@@ -38,9 +38,9 @@ public class ProjectResponse
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
     public string? Phone { get; set; }
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public float? Area { get; set; }
+    public double? Area { get; set; }
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public float? Depth { get; set; }
+    public double? Depth { get; set; }
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
     public string? Status { get; set; }
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
